refactor helper
